Add Check for Updates button in the Help menu
