+++ core/src/main/resources/l10n/messages.properties
@@ -1899,6 +1899,8 @@ main.menu.help.bugReport = Bug report
 main.menu.help.bugReport.desc = Information about reporting bugs in OpenRocket
 main.menu.help.debugLog = Debug log
 main.menu.help.debugLog.desc = View the OpenRocket debug log
+main.menu.help.checkForUpdates = Check for updates
+main.menu.help.checkForUpdates.desc = Check for updates to OpenRocket
 main.menu.help.about = About
 main.menu.help.about.desc = Copyright details about OpenRocket
